fix audio no sound, add missing VideoFrame timestamp
move get_time to hbb_common
